orm_query_tricks/requirements.txt
Django==2.2.2
django-debug-toolbar==3.2.1
djangorestframework==3.12.4
Faker==8.10.0
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
